requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
distro==1.9.0
lxml==6.0.4
flatbuffers==25.12.19
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
markdown-it-py==3.0.0
markdown_pdf==1.13.1
msal==1.35.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
onnxruntime==1.24.4
openai==2.30.0
packaging==26.0
plantuml==0.3.0
python-docx==1.2.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
pymupdf-layout==1.27.2.2
pymupdf4llm==1.27.2.2
pyparsing==3.3.2
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.0
ruff==0.15.8
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
sentencepiece==0.2.1
tabulate==0.10.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
